store local preference as long to support uint32 max value (#2818)

+++ projects/batfish/src/main/java/org/batfish/dataplane/protocols/BgpProtocolHelper.java
@@ -172,7 +172,7 @@ public class BgpProtocolHelper {
 
     // Outgoing nextHopIp & localPreference
     Ip nextHopIp;
-    int localPreference;
+    long localPreference;
     if (sessionProperties.isEbgp() || !remoteRouteIsBgp) {
       nextHopIp = fromNeighbor.getLocalIp();
       localPreference = BgpRoute.DEFAULT_LOCAL_PREFERENCE;
